pyproject.toml
[tool.poetry]
name = "etl_eco_bikes"
version = "0.1.0"
description = ""
authors = ["RArielAlonso <rariel.alonso@gmail.com>"]
readme = "README.md"
packages = [
    { include = "etl_modules",from = "dags"},
    { include = "config", from = "dags"},
    { include = "utils", from = "dags"},
    { include = "tests", from = "dags"}
    ]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.1.1"
requests = "^2.31.0"
pyarrow = "^13.0.0"
fastparquet = "^2023.8.0"
sqlalchemy = "^2.0.22"
psycopg2-binary = "^2.9.9"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
python-dotenv = "^1.0.0"
google-cloud-bigquery = "^3.13.0"
pandas-gbq = "^0.19.2"
setuptools = "^68.2.0"
google-cloud-storage = "^2.13.0"
gcsfs = "^2023.10.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
